### 2022-09-23 #4777 @editorialbot
[PRE REVIEW]: Wakeflow: A python package for semi-analytic models of planetary wakes
**Submitting author:** @TomHilder (<a href="http://orcid.org/0000-0001-7641-5235">Thomas Hilder</a>)
**Repository:** https://github.com/TomHilder/wakeflow
**Branch with paper.md** (empty if default branch): master
**Version:** v1.2.0
