### 2022-10-11 bee42bd
Added option to use old velocity prescription
