### 2022-10-12 b6775db
Changed linear box size definition
Added possibility to have different values for the left, right, top and bottom edges of the linear box
