+++ src/wakeflow/wakeflow.py
@@ -159,8 +159,10 @@ class WakeflowModel():
         adiabatic_index       = 1.6666667     # adiabatic index
         damping_malpha        = 0.0           # artificial damping NOT IMPLEMENTED
         CFL                   = 0.5           # Courant stability factor (require <0.5)
-        scale_box             = 1.0           # linear box length scale factor in radial direction
-        scale_box_ang         = 1.0           # linear box length scale factor in angular direction
+        scale_box_l           = 1.0           # linear box length scale factor in radial direction (left)
+        scale_box_r           = 1.0           # linear box length scale factor in radial direction (right)
+        scale_box_ang_t       = 1.0           # linear box length scale factor in angular direction (top)
+        scale_box_ang_b       = 1.0           # linear box length scale factor in angular direction (bottom)
         tf_fac                = 1.0           # scale factor for t coordinate where self-similar solution is used
         show_teta_debug_plots = False         # show (t,eta,chi) space developer plots
         box_warp              = True          # interpret y coordinate of linear regime as arc length, or truly vertical? True (default) for former
