### 2022-10-13 65d1a61
Changed linear box size definition
Added possibility to have different values for the left, right, top and bottom edges of the linear box (forgot to update the square box)
