### 2022-10-13 c38e65f
Changed linear box size definition
Added possibility to have different values for the left, right, top and bottom edges of the linear box (fixed the max function)
