### 2022-10-13 20a98f5
Changed linear box size definition
Added possibility to have different values for the left, right, top and bottom edges of the linear box
