### 2022-10-14 e42f06e
Changed linear box size definition
Forgot to update _cut_box_square
