### 2022-10-14 7e22506
Changed linear box size definition
Forgot to update the standard .yaml
