+++ src/wakeflow/data/default_config.yaml
@@ -63,8 +63,10 @@ damping_malpha: 0.0                          # artificial damping
 
 #numerical:
 CFL: 0.5                                     # CFL safety factor for Godunov scheme that solves Burger's eqn
-scale_box: 1.0                               # scale factor for linear regime box side length    -- default is 1
-scale_box_ang: 1.0                           # scale factor for linear regime box angular extent -- default is 1
+scale_box_l: 1.0                               # scale factor for linear regime box side length (left side)      -- default is 1
+scale_box_r: 1.0                               # scale factor for linear regime box side length (right side)     -- default is 1
+scale_box_ang_b: 1.0                           # scale factor for linear regime box angular extent (bottom side) -- default is 1
+scale_box_ang_t: 1.0                           # scale factor for linear regime box angular extent (top side)    -- default is 1
 tf_fac: 1.0                                  # factor to multiply the final time Burger's equation is evolved to -- default is 1
 
 # developer
