### 2022-10-14 095b12b
Changed linear box size definition
updated tests
