### 2022-10-14 0b83027
Changed linear box size definition
updated test
