+++ src/wakeflow/data/default_config.yaml
@@ -72,5 +72,5 @@ tf_fac: 1.0                                  # factor to multiply the final time
 # developer
 show_teta_debug_plots: False                 # show t, eta space plots (including putting the linear solution into t,eta)
 box_warp: True                               # whether or not the linear box warps with the disk curvature (bool) -- default is True
-use_box_IC: False                            # whether or not to use the side of the box as seen in the results as the initial condition 
-                                             # for the non-linear evolution (bool) -- default is False, DO NOT CHANGE
+use_box_IC: False                            # whether or not to use the side of the box as seen in the results as the initial condition for the non-linear evolution (bool) -- default is False, DO NOT CHANGE
+use_old_vel: False                           # use old approximated formulas for u pert
