### 2022-10-18 #4863 @editorialbot
[REVIEW]: Wakeflow: A python package for semi-analytic models of planetary wakes
**Submitting author:** @TomHilder (<a href="http://orcid.org/0000-0001-7641-5235">Thomas Hilder</a>)
**Repository:** https://github.com/TomHilder/wakeflow
**Branch with paper.md** (empty if default branch): master
**Version:** 1.3.1
