### 2022-11-16 TomHilder/wakeflow#10 @andizq
[bug] Automated tests
The automated tests fail if the standard [`pytest`](https://docs.pytest.org/en/7.2.x/) is used. Your tests require the [`pytest-cov`](https://pytest-cov.readthedocs.io/en/latest/readme.html#installation) plugin to support coverage reports. Please edit readme and documentation accordingly or modify functionality.

This is part of the reviewing process for JOSS openjournals/joss-reviews#4863
