### 2022-11-16 TomHilder/wakeflow#11 @andizq
[documentation] Various suggestions
- Missing density units in 'Wakeflow Parameters' section.
- In the tutorial consider using `ax.contourf(X[:,i,:], Y[:,i,:], rho[:,i,:], levels=256)` instead of `ax.pcolormesh` to avoid confusion with the order of x, y axes and the need to use the transpose of `rho`. Note that pcolormesh is intended for irregular rectangular grids and thus not really needed here.
- In the tutorial please consider adding a few words explaining why having a planet-to-thermal mass ratio > 1 may lead to unreliable results in the analytical framework.
- In the 'Disk Structure' section you should explicitly mention that the additional factor in the rotation velocity is due to the presence of radial pressure forces and provide actual link(s) to the reference(s) e.g. [Nelson et al. 2013](https://ui.adsabs.harvard.edu/abs/2013MNRAS.435.2610N/abstract).
- A few more details should be added as to what the user is seeing as output. For instance, in the tutorial you show an example of how to plot the resulting density field at different scale heights which may give the wrong impression that the velocity field can also vary as function of height (which is not accounted by the analytical theory in which the code is based). In this case please remind the user that the output velocity field is computed for the disc midplane only.

https://github.com/openjournals/joss-reviews/issues/4863
