### 2022-11-16 TomHilder/wakeflow#12 @andizq
[enhancement] Planet location
Please consider having the position angle of the planet (PAp) as an optional argument in wakeflow too (i.e. not only for the interface with MCFOST). This will make it easier to compare output velocities to actual observations and will also facilitate the link between wakeflow and other codes in the community such as eddy or discminer.

https://github.com/openjournals/joss-reviews/issues/4863
