### 2022-11-16 TomHilder/wakeflow#13 @andizq
[question] Grid vertical extent
Why is the maximum vertical extent of the grid fixed to one pressure scale height (at R_out)? To date the majority of velocity kinks reported in the literature have been identified in 12CO, which is normally between 2-4H over the disc midplane, and at large radial separations. I would then suggest you to make `z_max` an optional argument too with a default value of e.g. 3H at R_out.

https://github.com/openjournals/joss-reviews/issues/4863
