### 2022-11-16 #4863 @richteague
Hi @tomhilder! I've managed to successfully install and run `Wakeflow`. It's a very nice package! Just a few things which I've noticed as I've been working my way through the documentation:

1. The testing requires `pytest-cov`, which @andizq has already mentioned https://github.com/TomHilder/wakeflow/issues/10.

2. In the Quickstart Tutorial documentation, there's a few typos when talking about the results (just below cell 6). You've used `delta_rho` where I think you mean `delta_vphi` or similar. Further down in this document you also have a comment in cell 9 which describes a 1Mj example, but this is a 0.5Mj example.

3. In the Advanced Configuration section, I wonder whether the `dict()` syntax would be more useful than the `{}` format for writing the dictionary in cell 7 as this avoids having to explain the need to make the key a string.
