### 2022-11-16 #4863 @richteague
To add to these comments, for the paper I'd recommend to expand the Summary section to be a bit more explicit about what properties about the wake you're able to extract (e.g., the velocity components and rho), and then say that this is how we can connect with observations as those are what are being observed.
