### 2023-02-07 6173368
readme updated to include pytest-cov as requirement
