+++ README.md
@@ -109,6 +109,7 @@ Python packages:
 * `pyyaml`
 * `tqdm`
 * `pytest` (optional)
+* `pytest-cov` (optional)
 * [`pymcfost`](https://github.com/cpinte/pymcfost) (optional, only if interfacing with [MCFOST](https://github.com/cpinte/mcfost))
 
 If you install `wakeflow` using `pip` then the required dependencies will be automatically installed.
@@ -126,7 +127,7 @@ Please refer to the [Quickstart tutorial](https://wakeflow.readthedocs.io/en/lat
 <!-- USAGE EXAMPLES -->
 ## Testing
 
-`wakeflow` is automatically unit-tested on Github using Actions and [`tox`](https://github.com/tox-dev/tox). If you have installed `wakeflow` from source, you may run a local test on your machine provided that you have `pytest` installed. Simply navigate to your installation directory and run:
+`wakeflow` is automatically unit-tested on Github using Actions and [`tox`](https://github.com/tox-dev/tox). If you have installed `wakeflow` from source, you may run a local test on your machine provided that you have `pytest` and `pytest-cov` installed. Simply navigate to your installation directory and run:
 ```sh
 pytest
 ```
