### 2023-02-07 8f9ba33
removed python 3.6 from testing
