+++ tox.ini
@@ -2,11 +2,12 @@
 
 [tox]
 minversion = 3.8.0
-envlist = python3.7, python3.9, python3.10
+envlist = python3.6, python3.7, python3.9, python3.10
 isolated_build = true
 
 [gh-actions]
 python =
+    3.6: python3.6
     3.7: python3.7
     3.8: python3.8
     3.9: python3.9
