### 2023-02-07 b8629b3
Update Tests.yml
attempting to fix issue where python 3.6 can't be installed on latex ubuntu version for testing
