+++ .github/workflows/Tests.yml
@@ -9,7 +9,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [ubuntu-latest, windows-latest, macos-latest]
+        os: [ubuntu-20, windows-latest, macos-latest]
         python-version: ['3.6', '3.7', '3.8', '3.9', '3.10']
 
     steps:
