### 2023-02-07 TomHilder/wakeflow#10 @TomHilder
The documentation has been updated accordingly to make it clear that `pytest-cov` is also required for testing. Thanks!
