### 2023-02-08 9f47c5d
changed tests.yml to account for python3.6 on ubuntu
