+++ README.md
@@ -218,7 +218,7 @@ Additional acknowledgements:
 [license-url]: https://opensource.org/licenses/MIT
 [JOSS-shield]: https://img.shields.io/badge/JOSS-under_review-yellow?style=flat-square
 [JOSS-url]: https://joss.theoj.org/papers/22d35f9b0bb35a8df4dab85b0b6f4eb7
-[test-status-shield]: https://img.shields.io/github/workflow/status/TomHilder/wakeflow/Tests?label=tests&style=flat-square
+[test-status-shield]: https://img.shields.io/github/actions/workflow/status/TomHilder/wakeflow/test.yml?branch=main&label=tests&style=flat-square
 [test-status-url]: https://github.com/TomHilder/wakeflow/actions/workflows/Tests.yml
 [docs-status-shield]: https://readthedocs.org/projects/wakeflow/badge/?version=latest&style=flat-square
 [docs-status-url]: https://wakeflow.readthedocs.io/en/latest/?badge=latest
