+++ docs/wakeflow_parameters.rst
@@ -32,7 +32,7 @@ Model Parameters
 +-------------------------+------------------------------+---------------------------+-----------------------------------------------------------------------------------------------+
 | ``hr``                  | ``float``                    | ``0.1``                   | The disk aspect ratio at the reference radius.                                                |
 +-------------------------+------------------------------+---------------------------+-----------------------------------------------------------------------------------------------+
-| ``dens_ref``            | ``float``                    | ``1.0``                   | The gas density at the reference radius.                                                      |
+| ``dens_ref``            | ``float``                    | ``1.0``                   | The gas density at the reference radius in g/cm^3                                                      |
 +-------------------------+------------------------------+---------------------------+-----------------------------------------------------------------------------------------------+
 | ``cw_rotation``         | ``bool``                     | ``False``                 | Does the disk rotate clockwise?                                                               |
 +-------------------------+------------------------------+---------------------------+-----------------------------------------------------------------------------------------------+
