### 2023-02-09 TomHilder/wakeflow#11 @TomHilder
Thanks for the detailed feedback!

- Density units have been added as g/cm^3. The reason it was missing in the first place is because it is actually arbitrary. The calculation involves finding delta rho / rho, which is then scaled by the background density, which is itself scaled by the density reference parameter. The units of your result are therefore just whatever units you chose for the reference. Given that this is a detail, a bit confusing, and not so important, I've added units to the docs.
- I have indeed swapped `x, y, rho[:i,:].transpose()` for `X[:,i,:], Y[:,i,:], rho[:i,:]` as suggested to avoid the need for the transpose.
- I kept pcolormesh but added the optional argument `shading="nearest"` so that it is clear that the provided grid pointes are not being interpreted as the corners of cells and instead are the centres. I understand that I do not have an irregular grid, but pcolormesh is in general a better function that pcolor. I have not opted to use contourf because I think that cell shading is more honest about the results than filled contours (which involves a lot of interpolation), and it is also hard to do a log colorbar with contourf (that I know of =, anyway).
- I added some notes on the assumption that the planet mass is less than the thermal mass, and added links to appropriate references for further reading.
- I updated the disk structure section to mention radial pressure forces, and made the citation a link.
- I overhauled the visualisation section of the quickstart tutorial to hopefully improve the clarity of what the results do and do not include, and I have added some text on the fact that the analytics is 2D and so there is no z-dependence of the perturbations.

@andizq if you are satisfied with these changes let me know and I will close the issue. Thanks!
