### 2023-02-09 980f404
added ability to change max height
