+++ src/wakeflow/data/default_config.yaml
@@ -12,6 +12,7 @@ r_inner: 100                                 # inner radius of disk in au (does 
 r_planet: 250                                # orbital radius of embedded planet in au
 r_ref: null                                  # reference radius R0 for H/R -- YOU MUST USE null INSTEAD OF None IN A YAML FILE
 r_c: 0                                       # critical radius for exponentially tapered density profile. set to 0 for power law
+z_max: 3                                     # vertical extent of disk in units of scale height at r_outer
 q: 0.25                                      # q index, defined as c_s \propto R^{-q}
 p: 1.0                                       # p index, defined as rho \propto R^{-p} or \propto R^{-p}*\exp{-R/r_c}^{2-p}
 hr: 0.10                                     # disc flaring H/R at R0
