### 2023-02-09 8acf7af
added z_max to parameters docs
