+++ docs/wakeflow_parameters.rst
@@ -26,6 +26,8 @@ Model Parameters
 +-------------------------+------------------------------+---------------------------+-----------------------------------------------------------------------------------------------+
 | ``r_c``                 | ``float``                    | ``0``                     | Critical radius in AU for exponentially tapered disk. Exponential taper not used if set to 0. |
 +-------------------------+------------------------------+---------------------------+-----------------------------------------------------------------------------------------------+
+| ``z_max``               | ``float``                    | ``3``                     | Height of the disk, in units of pressure scale height at ``r_outer``.                         |
++-------------------------+------------------------------+---------------------------+-----------------------------------------------------------------------------------------------+
 | ``q``                   | ``float``                    | ``0.25``                  | The sound speed is proportional to R^-q.                                                      |
 +-------------------------+------------------------------+---------------------------+-----------------------------------------------------------------------------------------------+
 | ``p``                   | ``float``                    | ``1.0``                   | The density is proportional to R^-p.                                                          |
