### 2023-02-09 TomHilder/wakeflow#13 @TomHilder
Yes this is a great point. The suggested feature has been implemented and documentation updated accordingly.

Previously this has not mattered for us as in practice I have only used the Cartesian and cylindrical grid types for 2D analysis, and any time we needed to care about line emission we have been interfacing with `MCFOST`, which chooses the maximum z for itself (and `wakeflow` uses the `MCFOST` grid).
