+++ src/wakeflow/non_linear_perts.py
@@ -499,7 +499,7 @@ class _NonLinearPerts():
 
             [x_grid, y_grid] = np.meshgrid(x, y)
             r_grid = np.sqrt(x_grid**2 + y_grid**2)
-            pphi_grid = np.arctan2(y_grid, x_grid)
+            pphi_grid = np.arctan2(y_grid, x_grid) - self.p.phi_planet
 
             tt = _t_vector(r_grid, Rp, hr, q, p)
 
@@ -548,7 +548,7 @@ class _NonLinearPerts():
         # if using a cylindrical grid
         else:
             r = self.g.r
-            phi = self.g.phi
+            phi = self.g.phi - self.p.phi_planet
 
             r_grid, pphi_grid = np.meshgrid(r, phi)
 
