+++ src/wakeflow/model_setup.py
@@ -57,9 +57,10 @@ class _Parameters(_Constants):
             self.m_planet       = None
             self.m_planet_array = list(config["m_planet"])
 
-        self.r_outer  = float(config["r_outer"])
-        self.r_inner  = float(config["r_inner"])
-        self.r_planet = float(config["r_planet"])
+        self.r_outer    = float(config["r_outer"])
+        self.r_inner    = float(config["r_inner"])
+        self.r_planet   = float(config["r_planet"])
+        self.phi_planet = float(config["phi_planet"])
         try:
             self.r_ref = float(config["r_ref"])
         except:
