+++ docs/wakeflow_parameters.rst
@@ -22,6 +22,8 @@ Model Parameters
 +-------------------------+------------------------------+---------------------------+-----------------------------------------------------------------------------------------------+
 | ``r_planet``            | ``float``                    | ``250``                   | Orbital radius of planet in AU.                                                               |
 +-------------------------+------------------------------+---------------------------+-----------------------------------------------------------------------------------------------+
+| ``phi_planet``          | ``float``                    | ``250``                   | Azimuthal position of the planet in radians.                                                  |
++-------------------------+------------------------------+---------------------------+-----------------------------------------------------------------------------------------------+
 | ``r_ref``               | ``float``                    | ``r_planet``              | Reference radius in AU, where ``hr`` and ``dens_ref`` are defined.                            |
 +-------------------------+------------------------------+---------------------------+-----------------------------------------------------------------------------------------------+
 | ``r_c``                 | ``float``                    | ``0``                     | Critical radius in AU for exponentially tapered disk. Exponential taper not used if set to 0. |
