### 2023-02-10 TomHilder/wakeflow#12 @TomHilder
Hi @andizq, thanks for the suggestion! I have implemented this with the parameter `phi_planet`. Cheers!
