### 2023-02-10 8a08ec0
updated yaml reading to close file properly
