### 2023-02-10 #4863 @TomHilder
Hi @richteague , thank you for the feedback!

1.  The documentation has been updated to include `pytest-cov` as a requirement for testing.
2. The mentioned typos in the quickstart tutorial have been fixed.
3. This is a good suggestion and I have changed the tutorial to use the `dict()` syntax.

I am hoping to address your comments in regards to the paper early next week. Thanks!
