### 2023-02-10 #4863 @TomHilder
Hi @andizq , thank you also for the feedback! I have implemented all of your suggested changes, please see the issue threads for more details. I did stick with `pcolormesh` over `contourf` in the quickstart tutorial, and I have provided some justification in TomHilder/wakeflow#11 .
