### 2023-02-10 4a35c32
fixed default value of phi_planet
