+++ docs/wakeflow_parameters.rst
@@ -22,7 +22,7 @@ Model Parameters
 +-------------------------+------------------------------+---------------------------+-----------------------------------------------------------------------------------------------+
 | ``r_planet``            | ``float``                    | ``250``                   | Orbital radius of planet in AU.                                                               |
 +-------------------------+------------------------------+---------------------------+-----------------------------------------------------------------------------------------------+
-| ``phi_planet``          | ``float``                    | ``250``                   | Azimuthal position of the planet in radians.                                                  |
+| ``phi_planet``          | ``float``                    | ``0``                     | Azimuthal position of the planet in radians.                                                  |
 +-------------------------+------------------------------+---------------------------+-----------------------------------------------------------------------------------------------+
 | ``r_ref``               | ``float``                    | ``r_planet``              | Reference radius in AU, where ``hr`` and ``dens_ref`` are defined.                            |
 +-------------------------+------------------------------+---------------------------+-----------------------------------------------------------------------------------------------+
