### 2023-02-19 #4863 @TomHilder
Hi @richteague , I have updated the paper based on your feedback to be a bit more specific about what `wakeflow` calculates and how these can be connected to observations.

The summary now reads:

`Wakeflow` is a Python package for generating semi-analytic models of the perturbations
induced by planets embedded in gaseous circumstellar disks. These perturbations
take the form of a spiral shock wave [@Ogilvie:2002], and are often called
a "planet wake" in analogy with that produced by a boat in a lake.
Using `Wakeflow`, users may calculate the perturbed density and velocity fields of the gas in the disk.
These may be used with radiation transfer codes to generate synthetic observations
that map both the gas distribution and the gas kinematics.
Comparison with real observations, such as from molecular line emission taken with the
Attacama Large Millimetre Array, allows researchers to infer the properties of potential planets
as well as the disk itself.

Let me know if you think it needs to be updated further, I wasn't sure how specific to be when referring to observations.

I think that I have now addressed all the feedback that I received 😄
