### 2023-02-22 #4863 @andizq
Hi @TomHilder, thank you for addressing the feedback and suggestions in such a positive way! I am happy with (and have tested) the new additions and therefore recommend your work for publication as long as @richteague is also happy with the modifications. It is a very nice, easy-to-use tool that will facilitate the analysis and prediction of kinematic observables resulting from planet-disc interaction!
