### 2023-02-23 27b8623
Update azimuthal velocity with density tapering
Update grid.py. For reference Lodato et al. 2022 eq.12
