+++ src/wakeflow/grid.py
@@ -284,9 +284,14 @@ class _Grid:
         hrf = hr * (r / r_ref) ** (0.5 - q)
 
         # pressure and gravity height dependent correction
-        corr = np.sqrt(
-            (-1* (p + 2 * q) * hrf**2) + (1 - 2 * q) + (2 * q * r / np.sqrt(r**2 + z**2))
-        )
+        if r_c == 0: # no taper
+            corr = np.sqrt(
+                (-1* (p + 2 * q) * hrf**2) + (1 - 2 * q) + (2 * q * r / np.sqrt(r**2 + z**2))
+                )
+        else: #taper
+            corr = np.sqrt(
+                (-1* ((p + 2 * q) + (2 - p + 3/2 - q))* hrf**2) + (1 - 2 * q) + (2 * q * r / np.sqrt(r**2 + z**2))
+                )
 
         # perform correction
         self.v_phi *= corr
