### 2023-02-23 55917f3
Update grid.py
fixed an error
