### 2023-02-25 #4863 @dfm
@richteague, @andizq — Thank you for your thorough and constructive reviews! I really appreciate the time you volunteered to this process!! 🎉 🎉

@TomHilder — I believe that @editorialbot is right about that suggested DOI above ☝️. Can you update the reference for `Teague:2022` to the published version and then run the following steps:

1. Comment `@editorialbot generate pdf`, and take one last read through the manuscript to make sure that you're happy with it (it's harder to make changes later!), especially the author names and affiliations. I've taken a pass and it looks good to me!
