### 2023-02-27 dd17d12
added box blurring/weighting
