+++ paper/paper.bib
@@ -144,18 +144,21 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{Teague:2022,
-       author = {{Teague}, R. and {Bae}, J. and {Andrews}, S.~M. and {Benisty}, M. and {Bergin}, E.~A. and {Facchini}, S. and {Huang}, J. and {Longarini}, C. and {Wilner}, D.},
+       author = {{Teague}, Richard and {Bae}, Jaehan and {Andrews}, Sean M. and {Benisty}, Myriam and {Bergin}, Edwin A. and {Facchini}, Stefano and {Huang}, Jane and {Longarini}, Cristiano and {Wilner}, David},
         title = "{Mapping the Complex Kinematic Substructure in the TW Hya Disk}",
-      journal = {arXiv e-prints},
-     keywords = {Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
+      journal = {\apj},
+     keywords = {Protoplanetary disks, Exoplanet formation, Radio interferometry, 1300, 492, 1346, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
          year = 2022,
-        month = aug,
-          eid = {arXiv:2208.04837},
-        pages = {arXiv:2208.04837},
+        month = sep,
+       volume = {936},
+       number = {2},
+          eid = {163},
+        pages = {163},
+          doi = {10.3847/1538-4357/ac88ca},
 archivePrefix = {arXiv},
        eprint = {2208.04837},
  primaryClass = {astro-ph.EP},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2022arXiv220804837T},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2022ApJ...936..163T},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
