### 2023-02-27 #4863 @TomHilder
Hi @dfm! I have fixed the DOI and I am happy with the manuscript.

The version number is 1.3.1

The DOI for the Zenodo archive is 10.5281/zenodo.7679415

Thanks!
