+++ paper/paper.bib
@@ -146,7 +146,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Teague:2022,
        author = {{Teague}, Richard and {Bae}, Jaehan and {Andrews}, Sean M. and {Benisty}, Myriam and {Bergin}, Edwin A. and {Facchini}, Stefano and {Huang}, Jane and {Longarini}, Cristiano and {Wilner}, David},
         title = "{Mapping the Complex Kinematic Substructure in the TW Hya Disk}",
-      journal = {\apj},
+      journal = {The Astrophysical Journal},
      keywords = {Protoplanetary disks, Exoplanet formation, Radio interferometry, 1300, 492, 1346, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
          year = 2022,
         month = sep,
