### 2023-02-27 467f8e3
Add grid extension for rotation
Extend the current grid using the same spacing to account for corners during interpolation on rotated grid
