+++ src/wakeflow/grid.py
@@ -349,20 +349,21 @@ class _Grid:
             unperturbed density from Grid object where make_keplerian_disk has been used.
         """
 
-        # box size (in units of Hill radius), note for conversions that self.p.l = 1 Hill radius in cgs
-        x_box_size_l = 2 * self.p.scale_box_l
-        x_box_size_r = 2 * self.p.scale_box_r
-        y_box_size_t = 2 * self.p.scale_box_ang_t
-        y_box_size_b = 2 * self.p.scale_box_ang_b
+        # get linear solution           
+        lp = LinearPerts
 
-        min_r = self.p.r_planet - x_box_size_l * self.p.l
-        max_r = self.p.r_planet + x_box_size_r * self.p.l
-
-        min_y = -y_box_size_b * self.p.l
-        max_y =  y_box_size_t * self.p.l
-
-        max_phi =  np.pi / 2
-        min_phi = -np.pi / 2
+        # segment radial size (in units of Hill radius), note for conversions that self.p.l = 1 Hill radius in cgs. bsl/r stands for box_size_left/right
+        r_bsl = lp.x_box_l
+        r_bsr = lp.x_box_r
+        # segment azimuthal size (in units of \pi). bst/b stands for box_size_top/bottom
+        phi_bst = lp.x_box_t / 2
+        phi_bsb = lp.x_box_b / 2
+        # segment radial extrema for masking
+        min_r = self.p.r_planet - r_bsl * self.p.l
+        max_r = self.p.r_planet + r_bsr * self.p.l
+        # segment azimuthal extrema for masking
+        max_phi =  phi_bst * np.pi 
+        min_phi = -phi_bst * np.pi 
 
         # find (phi, r) grid for either Cartesian or Cylindrical global grid
         if self.info["Type"] == "cartesian":
@@ -373,43 +374,19 @@ class _Grid:
 
         # new PHI grid to use (-pi,pi) instead of (0,2pi), where values are swapped in place, also ditch z coordinate
         # also construct a mask that contains 0 outside linear annulus and 1 inside
-        PHI_new     = np.zeros((PHI.shape[0],PHI.shape[2]))
-        Y_new       = np.zeros((PHI.shape[0],PHI.shape[2]))
-        linear_mask = np.zeros((PHI.shape[0],PHI.shape[2]))
-
-        R_new = R[:,0,:]
-
-        for i in range(PHI.shape[0]):
-            for j in range(PHI.shape[2]):
-
-                # transforming phi coordinate in place
-                if PHI[i,0,j] > np.pi:
-                    PHI_new[i,j] = PHI[i,0,j] - 2*np.pi
-                else:
-                    PHI_new[i,j] = PHI[i,0,j]
-
-                Y_new[i,j] = R[i,0,j] * np.sin(PHI[i,0,j])
-
-                # constructing mask
-                if PHI_new[i,j] > min_phi and PHI_new[i,j] < max_phi \
-                    and Y_new[i,j] > min_y and Y_new[i,j] < max_y \
-                    and R_new[i,j] > min_r and R_new[i,j] < max_r:
-                    linear_mask[i,j] = 1
-                else:
-                    linear_mask[i,j] = 0
-
-        # get linear solution           
-        lp = LinearPerts
+        R_new       = R[:,0,:]
+        PHI_new     = np.where(PHI[:,0,:]>np.pi, PHI[:,0,:] - 2*np.pi, PHI[:,0,:])
[…]
