+++ src/wakeflow/linear_perts.py
@@ -12,6 +12,7 @@ import sys, pkg_resources, tarfile
 import numpy                    as np
 import matplotlib.pyplot        as plt
 from scipy.interpolate      import RectBivariateSpline
+from .transformations       import _Eta_vector
 
 if TYPE_CHECKING:
     from .model_setup       import _Parameters
@@ -29,67 +30,208 @@ class _LinearPerts():
 
         # grab parameters object
         self.p = parameters
-
-        # get location of linear perturbations data files
-        pert_loc = pkg_resources.resource_filename('wakeflow', 'data/linear_perturbations.npy')
-        mesh_loc = pkg_resources.resource_filename('wakeflow', 'data/linear_perturbations_mesh.npy')
-
-        # read perturbations from files
-        try:
-            perts = np.load(pert_loc)
-            mesh  = np.load(mesh_loc)
-
-        # in the case files have not been extracted from tarballs yet, extract them
-        except FileNotFoundError:
-
-            # open tarballs
-            pert_tar = tarfile.open(f"{pert_loc}.tar.gz")
-            mesh_tar = tarfile.open(f"{mesh_loc}.tar.gz")
-
-            # extract npy files
-            loc = pkg_resources.resource_filename('wakeflow', 'data')
-            pert_tar.extractall(loc)
-            mesh_tar.extractall(loc)
-
-            # close tarballs
-            pert_tar.close()
-            mesh_tar.close()
-
-            # read perturbations from files
-            perts = np.load(pert_loc)
-            mesh  = np.load(mesh_loc)
-
-        # get perturbation arrays
-        self.pert_v_r   = perts[0]
-        self.pert_v_phi = perts[1]
-        self.pert_rho   = perts[2]
-
-        # grid
-        self.X = mesh[0]
-        self.Y = mesh[1]
-
-        # linear perturbations read in grid
-        x = self.X[0,:]
-        y = self.Y[:,0]
-
-        # define constants for linear perts
-        self.x_box_l = 2 * self.p.scale_box_l
-        self.x_box_r = 2 * self.p.scale_box_r
-        self.x_box_t = 2 * self.p.scale_box_ang_t
-        self.x_box_b = 2 * self.p.scale_box_ang_b
         
-        # cut square box grid in linear regime
-        self.x_cut = x[np.argmin(x < -self.x_box_l) : np.argmin(x < self.x_box_r) + 1]
-        self.y_cut = y[np.argmin(y < -self.x_box_b) : np.argmin(y < self.x_box_t) + 1]
+        # initialise perturbations properties
+        self.info = {
+            "Type": None,
+            "Grid": None,
+            "Size": [0, 0, 0]
+            }
+        
+        # Using linear perturbations computed following Bollati et al. 2021 with shearing sheet assumption
+        # DEPRECATED: THIS WILL PROBABLY LEAD TO WRONG RESULTS
+        if self.p.lin_type == "shearing_sheet":
+            # get location of linear perturbations data files
+            pert_loc = pkg_resources.resource_filename('wakeflow', 'data/linear_perturbations.npy')
+            mesh_loc = pkg_resources.resource_filename('wakeflow', 'data/linear_perturbations_mesh.npy')
+    
+            # read perturbations from files
+            try:
+                perts = np.load(pert_loc)
+                mesh  = np.load(mesh_loc)
+    
+            # in the case files have not been extracted from tarballs yet, extract them
+            except FileNotFoundError:
+    
+                # open tarballs
+                pert_tar = tarfile.open(f"{pert_loc}.tar.gz")
+                mesh_tar = tarfile.open(f"{mesh_loc}.tar.gz")
+    
+                # extract npy files
[…]
