+++ src/wakeflow/non_linear_perts.py
@@ -15,7 +15,7 @@ from scipy.interpolate      import griddata
 from copy                   import copy
 from tqdm                   import tqdm
 from .burgers               import _solve_burgers
-from .transformations       import _Eta, _t, _get_chi, _get_dens_vel, _plot_r_t, _t_vector, _get_chi_vector
+from .transformations       import _Eta, _Eta_vector, _t, _get_chi, _get_dens_vel, _plot_r_t, _t_vector, _get_chi_vector, _g, _Lambda_fu, _Lambda_fv
 
 if TYPE_CHECKING:
     from .model_setup       import _Parameters
@@ -72,6 +72,14 @@ class _NonLinearPerts():
         y_max = 30
         y_rest = y[(y > -y_max) & (y < y_max)]
 
+        # plot for debugging
+        if False:
+            plt.contourf(x[index_inner:index_outer], y_rest, lp.pert_rho[(y > -y_max) & (y < y_max),index_inner:index_outer], cmap="RdBu", vmin=-1, vmax=1, levels=100)
+            plt.show()
+            plt.contourf(x[index_inner:index_outer], y_rest, lp.pert_rho[(y > -y_max) & (y < y_max),index_inner:index_outer], cmap="RdBu", vmin=-1, vmax=1, levels=100)
+            plt.ylim(-2.0,2.0)
+            plt.show()
+            
         # restrict inner and outer IC
         profile_rest_outer = profile_outer[(y > -y_max) & (y < y_max)]
         profile_rest_inner = profile_inner[(y > -y_max) & (y < y_max)]
@@ -89,7 +97,7 @@ class _NonLinearPerts():
         phi_IC_outer = y_IC_outer / self.p.r_planet
         phi_IC_inner = y_IC_inner / self.p.r_planet
 
-        if False:
+        if True:
             plt.plot(phi_IC_outer, profile_rest_outer, label="outer")
             plt.plot(phi_IC_inner, profile_rest_inner, label="inner")
             plt.legend(loc="best")
@@ -100,23 +108,36 @@ class _NonLinearPerts():
         eta_IC_inner = np.zeros(len(y_rest))
         t_IC_outer   = np.zeros(len(y_rest))
         t_IC_inner   = np.zeros(len(y_rest))
+        
+        eta_IC_outer = _Eta_vector(r_IC_outer, phi_IC_outer, self.p.r_planet, self.p.hr_planet, self.p.q, self.p.p, -1, self.p.m_planet, self.p.m_thermal)
+        eta_IC_inner = _Eta_vector(r_IC_inner, phi_IC_inner, self.p.r_planet, self.p.hr_planet, self.p.q, self.p.p, -1, self.p.m_planet, self.p.m_thermal)
 
+        # t for outer and inner
+        t_IC_outer = _t_vector(r_IC_outer, self.p.r_planet, self.p.hr_planet, self.p.q, self.p.p, self.p.m_planet, self.p.m_thermal)
+        t_IC_inner = _t_vector(r_IC_inner, self.p.r_planet, self.p.hr_planet, self.p.q, self.p.p, self.p.m_planet, self.p.m_thermal)
+
+
+        """
         # perform transformation
         for i in range(len(y_rest)):
 
             # eta for outer and inner
-            eta_IC_outer[i] = _Eta(r_IC_outer[i], phi_IC_outer[i], self.p.r_planet, self.p.hr_planet, self.p.q, -1)
-            eta_IC_inner[i] = _Eta(r_IC_inner[i], phi_IC_inner[i], self.p.r_planet, self.p.hr_planet, self.p.q, -1)
[…]
