+++ src/wakeflow/transformations.py
@@ -15,18 +15,20 @@ from scipy.interpolate      import RectBivariateSpline
 # TODO: update code to be faster and more efficient, most of the code in this file is still leftover from the old version (Analytical Kinks)
 
 # wake shape
-def _phi_wake(r, Rp, hr, q, cw): 
+def _phi_wake(r, Rp, hr, q, p, cw, m_p, m_th): 
     """Eq. (4) Bollati et al. 2021
     """
     rr = r / Rp
-    return -cw * np.sign(r - Rp) * (1 / hr) * (rr**(q - 0.5) / (q - 0.5) - rr**(q + 1) / (q + 1) - 3 / ((2 * q - 1) * (q + 1)))
+    phi_l = np.sign(r - Rp) * (1 / hr) * (rr**(q - 0.5) / (q - 0.5) - rr**(q + 1) / (q + 1) - 3 / ((2 * q - 1) * (q + 1)))
+    phi_nl = 0#np.sign(r - Rp) * 0.05 * hr * np.sqrt(_t_vector(rr, Rp, hr, q, p, m_p, m_th))
+    return -cw * (phi_l + phi_nl)
 
 # eta coordinate transformation
-def _Eta(r, phi, Rp, hr, q, cw):
+def _Eta(r, phi, Rp, hr, q, p, cw, m_p, m_th):
     """Eq. (14) Bollati et al. 2021
     """
     coeff    = 1.5 / hr
-    phi_w    = _mod2pi(_phi_wake(r, Rp, hr, q, cw))
+    phi_w    = _mod2pi(_phi_wake(r, Rp, hr, q, p, cw, m_p, m_th))
     deltaphi = phi - phi_w
 
     if deltaphi > np.pi:
@@ -37,7 +39,7 @@ def _Eta(r, phi, Rp, hr, q, cw):
 
     return coeff * deltaphi
 
-def _Eta_vector(r, phi, Rp, hr, q, cw):
+def _Eta_vector(r, phi, Rp, hr, q, p, cw, m_p, m_th):
     """Eq. (14) Bollati et al. 2021
 
     Vectorised version of _Eta.
@@ -45,7 +47,7 @@ def _Eta_vector(r, phi, Rp, hr, q, cw):
     modulus operators and constant offsets.
     """
     coeff    = 1.5 / hr
-    phi_w    = _phi_wake(r, Rp, hr, q, cw) % (2*np.pi)
+    phi_w    = _phi_wake(r, Rp, hr, q, p, cw, m_p, m_th) % (2*np.pi)
     deltaphi = (phi - phi_w + np.pi) % (2*np.pi) - np.pi
 
     return coeff * deltaphi
@@ -76,7 +78,7 @@ def _t_integrand(x, q, p):
 def _t_integral(up, q, p):
     return  quad(_t_integrand, 1, up, args=(q,p))[0]
 
-def _t_vector(rr, Rp, hr, q, p):
+def _t_vector(rr, Rp, hr, q, p, m_p, m_th):
     """Equation (43) Rafikov 2002    (Eq. 13 Bollati et al. 2021)
     This is a vectorised version of _t.
     _t computes an integral where the integrand is independent of the radius r.
@@ -138,16 +140,16 @@ def _t_vector(rr, Rp, hr, q, p):
     module_integral = np.zeros_like(sorted_results)
     module_integral[integral_bounds_sorted_indices] = np.abs(sorted_results)
 
-    coeff = 3 * hr**(-5 / 2) / (2**(5 / 4))
+    coeff = 3 * hr**(-5 / 2) / (2**(5 / 4)) * (m_p/m_th)
     return np.reshape(coeff * module_integral, shape)
 
 
 # t coordinate transformation
[…]
