### 2023-02-27 e57febf
Linear perturbation update & QoL & Bug fixes
- Simplified physical rescaling (done once in non_linear_perts.py)
- Eliminated eta extension of profile, initial profile taken on annulus now
- Implemented cyclic boundary conditions
- Changed limit for numerical computation, equal to the edge of the disc
- Implemented automatic break condition when reaching asymptotic limit (wip, to be tested)
- Added some nice debug plots
