+++ src/wakeflow/burgers.py
@@ -7,6 +7,8 @@ Contains the solve_burgers function used in the wake non-linear propagation calc
 
 import numpy                as np
 import matplotlib.pyplot    as plt
+import matplotlib.colors    as cl
+import matplotlib.cm        as cm
 
 # NOTE: contents are intended for internal use and should not be directly accessed by users
 
@@ -23,14 +25,15 @@ def _solve_burgers(
     linear_solution, 
     linear_t, 
     show_teta, 
-    tf_fac
+    tf_fac,
+    edge_limit
 ): 
     """Propagate the wake in (t,eta,chi) space by solving Eq. 10 from Bollati et al. 2021 using Godunov scheme.
     """
 
-    # Eq. (15) Bollati et al. 2021
-    profile = profile * (gamma + 1) * beta_p / 2**(3 / 4)
-    deta    = eta[1] - eta[0]
+    #Get eta step
+    deta = eta[1] - eta[0]
+    print(len(eta))
 
     # time required to develop N-wave for betap = 1
     tf_th = 300 
@@ -38,11 +41,8 @@ def _solve_burgers(
     # time required to display N-wave for generic betap, Eq. (39) Rafikov 2002
     tf  = tf_th / beta_p 
     tf *= tf_fac
-
-    # Eq. (18) Bollati et al. 2021
-    eta_min = -eta_tilde - np.sqrt(2 * C * tf) - 3 
-    eta_max = -eta_tilde + np.sqrt(2 * C * tf) + 3
-
+    
+    """
     # extend the profile domain due to profile broadening during evolution 
     extr = eta[-1]
 
@@ -65,7 +65,7 @@ def _solve_burgers(
 
         #if show_teta:
         #    linear_solution = np.insert(linear_solution, 0, 0, axis=0)
-
+    """
     # number of centers
     Neta = len(eta)
 
@@ -109,7 +109,7 @@ def _solve_burgers(
     # initialise
     lapsed_time = 0
     counter     = 0
-
+    """
     # time integration
     while lapsed_time < tf:
 
@@ -127,60 +127,180 @@ def _solve_burgers(
         F[1:Neta] = _NumericalFluxVector(solution[-1][0:-1], solution[-1][1:])
 
         # compute the left boundary flux
-        if solution[-1][0] < 0.0:
-            uL = 2.0 * solution[-1][0] - solution[-1][1]
-        else:
-            uL = solution[0][0]
-
+        uL   = solution[-1][-1]
         uR   = solution[-1][0]
         F[0] = _NumericalFluxVector(uL, uR)
 
         # compute the right boundary flux
-        if solution[-1][Neta - 1] > 0.0:
-            uR = 2.0 * solution[-1][Neta - 1] - solution[-1][Neta - 2]
-        else:
-            uR = solution[0][Neta - 1]
-
+        uR      = solution[-1][0]
         uL      = solution[-1][Neta - 1]
         F[Neta] = _NumericalFluxVector(uL, uR)
 
         solution.append(solution[-1][0:Neta] - dt / deta * (F[1:Neta+1] - F[0:Neta]))
 
+        # Eq. (18) Bollati et al. 2021
+        eta_minus = eta_tilde - np.sqrt(2 * C * tf)
+        eta_plus = eta_tilde + np.sqrt(2 * C * tf)
+
+        if lapsed_time >= edge_limit:
+            break
+    """
+    # time integration until boundary of the disc
+    print(edge_limit)
+    while lapsed_time <= edge_limit:
+
+        # increment
[…]
