### 2023-02-27 9dde002
mcmc update
Crude mcmc approach, need to write proper interface
