### 2023-02-27 ad1706b
mcmc update
Crude mcmc approach, need to write proper interface
