+++ src/wakeflow/grid.py
@@ -771,3 +771,10 @@ class _Grid:
         np.save(f"{savedir}/{label}_rho.npy", self.rho)
 
         print(f"{printed} saved to {savedir}")
+
+        # method to get the velocity perturbations to feed to the mcmc chain
+        def _get_velocity_perturbations(self) -> float:
+
+            return self.v_r, self.v_phi, self.X, self.Y
+
+
