+++ src/wakeflow/model_setup.py
@@ -101,6 +101,7 @@ class _Parameters(_Constants):
         self.save_total         = bool(config["save_total"])
         self.write_FITS         = bool(config["write_FITS"])
         self.dimensionless      = bool(config["dimensionless"])
+        self.mcmc               = bool(config["mcmc"])
 
         # mcfost parameters
         self.run_mcfost = bool (config["run_mcfost"])
@@ -234,10 +235,18 @@ class _Parameters(_Constants):
         print("Parameters Ok - continuing")
         return True
         
-        # check linear perturbations input from user
+        # warning for saving results in mcmc
+        if self.mcmc and self.save_perturbations:
+            print("WARNING: Do not save the perturbation fields when using wakeflow inside the mcmc chain")
+        
+        # warning for saving results in mcmc
+        if self.mcmc and self.save_total:
+            print("WARNING: Do not save the density and velocity fields when using wakeflow inside the mcmc chain")
+            
+        # check linear perturbations input from user    
         if self.lin_type != "shearing_sheet" and self.lin_type != "global" and self.lin_type != "simulation":
             raise Exception("Invalid linear perturbation type. Choose either global or simulation or shearing_sheet)")
-        
+            
         # Warning on shearing sheet approximation
         if self.lin_type == "shearing_sheet":
             print("WARNING: The shearing sheet approximation may be invalid for a given choice of parameters. This may lead to incorrect results")
