### 2023-02-28 #4863 @dfm
@xuanxu — Thanks for quickly fixing the acceptance pipeline!! 🚀

@richteague, @andizq — Many thanks for your reviews here! JOSS relies upon the volunteer effort of people like you and we simply wouldn't be able to do this without you!!
